backend/requirements.txt
django
pypdf
Pillow
pymupdf
pdf2docx
python-pptx
pikepdf
openpyxl
xlwt
img2pdf
xhtml2pdf
markdown
mammoth
xlrd
docx2python
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-allauth
dj-rest-auth
python-dotenv
olefile
redis
celery
django-redis
django-storages[boto3]
flower
